Add files via upload
HTML calculator is used for performing basic mathematical operations like Addition, subtraction, multiplication, and division.

To design the basic calculator, we will use HTML, CSS, and JavaScript. HTML is used to design the basic structure of the calculator. CSS styles are used to apply styles on the calculator and JavaScript is used to add the calculation functionality.


Approach:

1) Created the design with the HTML Table where firstis holding the input field with id=”result” and the rest are filled with input button.
2) With every click of the button, it displays the respective value of the button to the input field by using the function dis().
3) myFunction() is used to set the value pressed from the keyboard to the same input field.
4) Calculation of the numbers can be done by both the “Enter” key as well as the “=” button on the Calculator UI. The solve() function is evaluating the result with the math.evaluate() and display the final answer to the input field with id=”result”.
5) clr() function is also defined to clear the input field.